+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * the user's picture can be shown next to each post. It is also shown in the
     user's profile in the admin panel and in the user bar;
 * `User` entity has the `picture` virtual field, which contains the path of the
-    user's picture or, alternatively, a default image.
+    user's picture or, alternatively, a default image;
+* updated for `cakephp-thumber` 1.3.0.
 
 ### 2.22.5-RC2
 * `Page` and `Post` entities have the `plain_text` virtual field. `Photo` entity